Fix misconfiguration error in Redis dependency
See: https://github.com/andymccurdy/redis-py/issues/995.

Briefly, the library Redis, in versions lower than 3.1.0, can throw a KeyError
in certain occasions if the ConnectionPool is created without the host/port/db
explicitly stated, even though it is not required to do so. This creates a
fairly non-obvious error hidden in the application's logs. Version 3.1.0
corrects the issue.